backend/requirements.txt
# SnapLearn AI Backend Dependencies
# Install: python -m pip install -r requirements.txt
# Use the same interpreter as your venv (e.g. .venv\Scripts\python.exe on Windows).

# --- Core API ---
fastapi==0.115.6
uvicorn[standard]==0.32.1
# realtime (supabase) requires pydantic>=2.11.7; keep aligned with google-genai.
pydantic==2.12.4
pydantic-settings==2.10.1
python-multipart==0.0.20
python-dotenv==1.0.1
httpx==0.28.1
aiofiles==24.1.0
psutil==6.1.0

# --- Google Gemini ---
google-genai==1.73.1

# --- Database (optional; MemoryManager falls back to local JSON) ---
# 2.10.x pins httpx<0.28, which conflicts with google-genai (needs httpx>=0.28.1).
supabase==2.29.0

# --- Multimodal input (required for input_processor.py / main.py) ---
Pillow==11.0.0
SpeechRecognition==3.12.0
pydub==0.25.1
langdetect==1.0.9

# --- Manim / video pipeline ---
manim==0.18.1
numpy==2.2.1
matplotlib==3.10.0
# TTS for video narration (edge-tts first, gTTS fallback)
edge-tts==6.1.9
gTTS==2.5.0

# --- Optional vision extras (image preprocessing) ---
opencv-python==4.10.0.84

# --- Development ---
pytest==8.3.4
pytest-asyncio==0.25.0
black==24.10.0
flake8==7.1.1
